Moved to imgen, deleted polydisperse option
